Adds custom validator and removes obsolete validators
